refactor(crystallize): split crystallize.mjs along its own section boundaries
crystallize.mjs had grown to 2317 lines and eight in-flight items all held it
at once, so every change to the close gate, the apply path or the parking
logic serialized behind the same file. The author had already drawn the
boundaries as section comments; this turns those comments into files.

  scripts/lib/crystallize-args.mjs        argument parsing
  scripts/lib/crystallize-close-gate.mjs  the hard gate
  scripts/lib/crystallize-close-check.mjs the check path
  scripts/lib/crystallize-close-apply.mjs the apply path and the marker section
  scripts/lib/crystallize-helpers.mjs     shared helpers

crystallize.mjs goes from 2317 lines to 224. The move itself changes no
behavior, but four things had to move with it.

Making applySessionClose an export opened a door that did not exist before.
Approval belongs to a human and no hook may reach an apply path; a test scans
every hooks/*.mjs to hold that. While the function was private inside
crystallize.mjs a hook could not have imported it however hard it tried, so
the scan never had to name it. Now it can be imported, called, or reached by
spawning its CLI flag, and the scan covers all three.

The prompt-surface guard in tests/collect-changelog.test.mjs derives a
script's flags from that one file's source. With the parsing delegated to
crystallize-args.mjs it found zero flags and fired the canary its author
left for exactly this. It now follows a script's local imports one hop,
but only when the script itself parses nothing: a plain union would let
weekly-report.mjs inherit --max-age-days from session-audit.mjs and sweep
a document that passes it, which is the fail-open direction.

Relocating a private parseFrontmatter into scripts/lib/ collided with the
canonical one in lib/frontmatter.mjs, which is stricter and correct where
this one is not (a nested relations: block overwrites a top-level key here).
The divergence predates this change and its only consumer is an advisory
candidate list, so this renames the moved copy to parseFrontmatterLoose and
says in a comment which one new code should take.

Four cross-file references pointed at line numbers inside the old file.
They now name the function instead, so the next move does not break them
again.

The risk here is the ship surface rather than the logic: package.json's
files list names every scripts/ file one by one, so a new file that is not
listed simply never reaches an installed copy. All five are added there and
to the pack-surface snapshot.

applySessionClose and the marker section stayed together because the marker
code continues straight out of applySessionClose with no header of its own.
That leaves crystallize-close-apply.mjs at 1739 lines, 63% of it that one
function, and the items that used to queue behind the old file now queue
behind this one. The split is a prerequisite for separating them, not the
separation itself.
