+++ package.json
@@ -28,6 +28,11 @@
     "scripts/verify.mjs",
     "scripts/weekly-report.mjs",
     "scripts/lib/core-hooks.mjs",
+    "scripts/lib/crystallize-args.mjs",
+    "scripts/lib/crystallize-close-apply.mjs",
+    "scripts/lib/crystallize-close-check.mjs",
+    "scripts/lib/crystallize-close-gate.mjs",
+    "scripts/lib/crystallize-helpers.mjs",
     "scripts/lib/design-history-stale.mjs",
     "scripts/lib/extensions.mjs",
     "scripts/lib/failure-type.mjs",
